Ingest of unpacked SIP as requested by archives

+++ ContentBroker/src/main/java/de/uzk/hki/da/cb/UnpackAction.java
@@ -91,8 +91,12 @@ public class UnpackAction extends AbstractAction {
 
 	@Override
 	public boolean implementation() throws IOException{
+		long size = 0L;
+		if (sipContainerOnIngestAreaIsDir())
+		 size = FileUtils.sizeOfDirectory(sipContainerOnIngestArea());
+		else size = sipContainerOnIngestArea().length();
 		
-		if (!ingestGate.canHandle(sipContainerOnIngestArea().length())){
+		if (!ingestGate.canHandle(size)){
 //			JmsMessage jms = new JmsMessage(C.QUEUE_TO_CLIENT,C.QUEUE_TO_SERVER,o.getIdentifier() + " - Please check WorkArea space limitations: " + ingestGate.getFreeDiskSpacePercent() +" % free needed " );
 //			super.getJmsMessageServiceHandler().sendJMSMessage(jms);	
 			logger.warn("ResourceMonitor prevents further processing of package due to space limitations. Setting job back to start state.");
@@ -101,11 +105,14 @@ public class UnpackAction extends AbstractAction {
 		
 		
 		wa.ingestSIP(sipContainerOnIngestArea());
-		unpack(wa.sipFile());
-		wa.sipFile().delete();
-
-		
-
+		if (!sipContainerOnIngestAreaIsDir()) {
+			unpack(wa.sipFile());
+			expandDirInto();
+			wa.sipFile().delete();
+		} else {
+			moveSipDir();
+			expandDirInto();
+		}
 		
 		throwUserExceptionIfNotBagitConsistent();
 		throwUserExceptionIfDuplicatesExist();
@@ -113,14 +120,24 @@ public class UnpackAction extends AbstractAction {
 		
 		// Is the last step of action because it should only happen after validity has been proven. 
 		logger.info("Removing SIP from IngestArea");
+		if (!sipContainerOnIngestAreaIsDir()) {
 		sipContainerOnIngestArea().delete();
+		} else FileUtils.deleteDirectory(sipContainerOnIngestArea());
 		return true;
 	}	
 	
+	private void moveSipDir() throws IOException {
+		FileUtils.moveDirectoryToDirectory(wa.sipFile(), wa.objectPath().toFile(), true);
+	}
 	
 	private File sipContainerOnIngestArea() {
 		return sipContainerInIngestAreaPath().toFile();
 	}
+	
+	private boolean sipContainerOnIngestAreaIsDir() {
+		return sipContainerInIngestAreaPath().toFile().isDirectory();
+	}
+	
 	
 	private Path sipContainerInIngestAreaPath() {
 		return Path.make(
@@ -135,6 +152,7 @@ public class UnpackAction extends AbstractAction {
 	public void rollback() throws IOException {
 		
 		FileUtils.deleteDirectory(wa.objectPath().toFile());
+		if (!sipContainerOnIngestAreaIsDir())
 		wa.sipFile().delete();
 		
 		o.getLatestPackage().getFiles().clear();
@@ -249,40 +267,7 @@ public class UnpackAction extends AbstractAction {
 		return documentsToFiles;
 	}
 	
-	
-	
-	
-	
-	
-	
-	
-	
-	
-	
-	
-	
-	
-	/**
-	 * Creates a folder at targetFolderPath and expands the contents of sourceFilePath into it.
-	 * @param sourceFilePath
-	 * @param targetFolderPath
-	 * @throws RuntimeException if the folder at targetFolderPath already exists or the file at 
-	 * sourceFilePath doesn't exist or the archive couldn't be unpacked.
-	 */
-	private void unpack(File sourceFile){
-		
-		wa.objectPath().toFile().mkdir();
-		
-		
-		if (!sourceFile.exists())
-			throw new RuntimeException("container at "+ sourceFile + " doesn't exist");
-		
-		ArchiveBuilder builder = ArchiveBuilderFactory.getArchiveBuilderForFile(sourceFile);
-		try {
-			builder.unarchiveFolder(sourceFile, wa.objectPath().toFile());
-		} catch (Exception e) {
-			throw new RuntimeException("couldn't unpack archive", e);
-		}
+	private void expandDirInto() {
 
 		File[] files = wa.objectPath().toFile().listFiles();
 		if (files.length == 1) {
@@ -313,6 +298,29 @@ public class UnpackAction extends AbstractAction {
 				throw new RuntimeException("couldn't delete folder " + files[0].getAbsolutePath());
 			}
 		}		
+
+	}
+	/**
+	 * Creates a folder at targetFolderPath and expands the contents of sourceFilePath into it.
+	 * @param sourceFilePath
+	 * @param targetFolderPath
+	 * @throws RuntimeException if the folder at targetFolderPath already exists or the file at 
+	 * sourceFilePath doesn't exist or the archive couldn't be unpacked.
+	 */
+	private void unpack(File sourceFile){
+		
+		wa.objectPath().toFile().mkdir();
+		
+		
+		if (!sourceFile.exists())
+			throw new RuntimeException("container at "+ sourceFile + " doesn't exist");
+		
+		ArchiveBuilder builder = ArchiveBuilderFactory.getArchiveBuilderForFile(sourceFile);
+		try {
+			builder.unarchiveFolder(sourceFile, wa.objectPath().toFile());
+		} catch (Exception e) {
+			throw new RuntimeException("couldn't unpack archive", e);
+		}
 	}
